Move `import_object` to where it is used
